+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+- database: add GetAllManagedDatabases function
+
 ## [8.27.0]
 
 ### Added